cccam: removed recursive locks, changed locking behaviour loadbalancer: fixed lb-value webif: showing cccam/cccam ext mode
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@4258 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-stat.c
@@ -306,7 +306,7 @@ int get_best_reader(ECM_REQUEST *er)
 	int i;
 	uint8 result[CS_MAXREADER];
 	memset(result, 0, sizeof(result));
-	uint8 re[CS_MAXREADER];
+	int re[CS_MAXREADER];
 
 	//resulting values:
 	memset(re, 0, sizeof(re));
